pyproject.toml
[project]
name = "practice-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beanie>=2.0.1",
    "fastapi>=0.135.1",
    "motor>=3.7.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyjwt[crypto]>=2.12.1",
    "python-multipart>=0.0.22",
    "redis[asyncio]>=7.3.0",
    "slowapi>=0.1.9",
    "uvicorn>=0.42.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]
pythonpath = ["."]

[dependency-groups]
dev = [
    "fakeredis>=2.34.1",
    "httpx>=0.28.1",
    "mongomock-motor>=0.0.36",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pre-commit>=4.1.0",
    "ruff>=0.15.7",
]

[tool.ruff]
line-length = 100
exclude = [
    ".git",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    "venv",
]

[tool.ruff.lint]
# E: pycodestyle, F: Pyflakes, I: isort, B: flake8-bugbear
select = ["E", "F", "I", "B"]
ignore = []
